Squashed 'lib/LazyDiyDict/' changes from 75d7e77a..bb11fedf
bb11fedf a namespace scope variable had been declared and defined in the header file. Moved the definition to source file along with changing it's uses to actually rely on the variable and not a static typing of the code user.

git-subtree-dir: lib/LazyDiyDict
git-subtree-split: bb11fedf67f56430bda66070c564eab5c2792c68